moving service logic into service.go, parameter input into input.go for payment
